refactor: Swap the erroneous content of modules
