+++ ewm-main/src/main/java/ru/practicum/ewm/MainService.java
@@ -1,11 +1,11 @@
-package ru.practicum;
+package ru.practicum.ewm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class StatsService {
+public class MainService {
     public static void main(String[] args) {
-        SpringApplication.run(StatsService.class, args);
+        SpringApplication.run(MainService.class, args);
     }
 }
