+++ ewm-main/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 9090
+  port: 8080
 
 spring:
   datasource:
@@ -8,7 +8,6 @@ spring:
     username: ${app.db.user}
     password: ${app.db.password}
   jpa:
-    #    database: POSTGRESQL
     hibernate:
       ddl-auto: none
     properties:
@@ -20,13 +19,15 @@ spring:
       mode: always
 
 app:
-  name: ewm-stats
+  name: ewm-main
+  stats-server-url: http://localhost:9090
   db:
     host: localhost
     port: 5432
-    name: ewm-stats
+    name: ewm-main
     user: postgres
     password: 12345
   format:
     date: yyyy-MM-dd
     date-time: yyyy-MM-dd HH:mm:ss
+
