+++ ewm-stats/src/main/java/ru/practicum/ewm/StatsService.java
@@ -1,11 +1,11 @@
-package ru.practicum;
+package ru.practicum.ewm;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class EwmService {
+public class StatsService {
     public static void main(String[] args) {
-        SpringApplication.run(EwmService.class, args);
+        SpringApplication.run(StatsService.class, args);
     }
 }
