+++ ewm-stats/src/main/java/ru/practicum/ewm/stats/model/StatsDtoResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.stats.model;
+package ru.practicum.ewm.stats.model;
 
 import lombok.Value;
