+++ ewm-stats/src/main/java/ru/practicum/ewm/stats/storage/StatsRepository.java
@@ -1,0 +1,8 @@
+package ru.practicum.ewm.stats.storage;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.ewm.stats.model.Stats;
+
+public interface StatsRepository extends JpaRepository<Stats, Long> {
+
+}
