+++ ewm-stats/src/main/resources/schema.sql
